APK File
Everything works fine in the Unity editor and according to the code, but I'm constantly facing issues in the builds. A few hours ago, I couldn't even get a build, so I deleted the editor and reinstalled it, but these kinds of issues might have occurred due to the version difference.